+++ .github/workflows/start_mongo.sh
@@ -5,6 +5,12 @@ MONGODB=$1
 mongodb_dir=$(find ${PWD}/ -type d -name "mongodb-linux-x86_64*")
 
 mkdir $mongodb_dir/data
+
 $mongodb_dir/bin/mongod --dbpath $mongodb_dir/data --logpath $mongodb_dir/mongodb.log --fork --replSet mongoengine
-mongo --eval 'db.version();'    # Make sure mongo is awake
-mongo --eval "rs.initiate()"
+if (( $(echo "$MONGODB < 6.0" | bc -l) )); then
+mongo --quiet --eval 'db.runCommand("ping").ok'    # Make sure mongo is awake
+mongo --quiet --eval "rs.initiate()"
+else
+mongosh --quiet --eval 'db.runCommand("ping").ok'  # Make sure mongo is awake
+mongosh --quiet --eval "rs.initiate()"
+fi