+++ mongoengine/base/datastructures.py
@@ -113,7 +113,11 @@ class BaseList(list):
         BaseDocument = _import_class("BaseDocument")
 
         if isinstance(instance, BaseDocument):
-            self._instance = weakref.proxy(instance)
+            if isinstance(instance, weakref.ProxyTypes):
+                self._instance = instance
+            else:
+                self._instance = weakref.proxy(instance)
+
         self._name = name
         super().__init__(list_items)
 
@@ -186,10 +190,6 @@ class BaseList(list):
 
 
 class EmbeddedDocumentList(BaseList):
-    def __init__(self, list_items, instance, name):
-        super().__init__(list_items, instance, name)
-        self._instance = instance
-
     @classmethod
     def __match_all(cls, embedded_doc, kwargs):
         """Return True if a given embedded doc matches all the filter