+++ mongoengine/mongodb_support.py
@@ -1,6 +1,7 @@
 """
 Helper functions, constants, and types to aid with MongoDB version support
 """
+
 from mongoengine.connection import get_connection
 
 # Constant that can be used to compare the version retrieved with
@@ -9,6 +10,9 @@ MONGODB_34 = (3, 4)
 MONGODB_36 = (3, 6)
 MONGODB_42 = (4, 2)
 MONGODB_44 = (4, 4)
+MONGODB_50 = (5, 0)
+MONGODB_60 = (6, 0)
+MONGODB_70 = (7, 0)
 
 
 def get_mongodb_version():